feat: separate components into packages
