+++ packages/avatar/package.json
