+++ packages/avatar/stories/Avatar.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { Meta } from '@storybook/react/types-6-0';
-import Avatar, { AvatarProps } from '../src/components/Avatar';
+import { Meta } from '@storybook/react';
+import Avatar, { AvatarProps } from '../src/index';
 
 export default {
   title: 'Components/Avatar',
@@ -28,14 +28,14 @@ export const Avatars = (args: AvatarProps) => (
   <div className="flex space-x-4">
     <Avatar
       src="https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80"
-      alt="User 1"
       {...args}
+      alt="User 1"
     />
     <Avatar
       src="https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80"
-      alt="User 2"
       {...args}
+      alt="User 2"
     />
-    <Avatar showPlaceholder={true} alt="User 2" {...args} />
+    <Avatar showPlaceholder={true} {...args} alt="User 2"  />
   </div>
 );
