+++ packages/avatar/tsconfig.json
@@ -1,0 +1,4 @@
+{
+    "extends": ["../../tsconfig.json"],
+    "include": ["src"]
+}
