--- src/index.tsx
@@ -1,3 +1,0 @@
-import './tailwind.css';
-
-export * from './components/Button';
